Fixed some integer conversion warnings.
Warnings present when compiling with clang-cl 11.0

+++ codemp/game/g_log.c
@@ -358,25 +358,25 @@ void G_LogWeaponOutput(void)
 	Q_strncpyz(mapname, Info_ValueForKey( info, "mapname" ), sizeof(mapname));
 
 	Com_sprintf(string, sizeof(string), "\n\n\nLevel:\t%s\n\n\n", mapname);
-	trap->FS_Write( string, strlen( string ), weaponfile);
+	trap->FS_Write( string, (int)strlen( string ), weaponfile);
 
 
 	// Combat data per character
 
 	// Start with Pickups per character
 	Com_sprintf(string, sizeof(string), "Weapon Pickups per Player:\n\n");
-	trap->FS_Write( string, strlen( string ), weaponfile);
+	trap->FS_Write( string, (int)strlen( string ), weaponfile);
 
 	Com_sprintf(string, sizeof(string), "Player");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 	for (j=0; j<WP_NUM_WEAPONS; j++)
 	{
 		Com_sprintf(string, sizeof(string), "\t%s", weaponNameFromIndex[j]);
-		trap->FS_Write(string, strlen(string), weaponfile);
+		trap->FS_Write(string, (int)strlen(string), weaponfile);
 	}
 	Com_sprintf(string, sizeof(string), "\n");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 	// Cycle through each player, give their name and the number of times they picked up each weapon.
 	for (i=0; i<MAX_CLIENTS; i++)
@@ -391,47 +391,47 @@ void G_LogWeaponOutput(void)
 			{
 				nameptr = unknownname;
 			}
-			trap->FS_Write(nameptr, strlen(nameptr), weaponfile);
+			trap->FS_Write(nameptr, (int)strlen(nameptr), weaponfile);
 
 			for (j=0;j<WP_NUM_WEAPONS;j++)
 			{
 				Com_sprintf(string, sizeof(string), "\t%d", G_WeaponLogPickups[i][j]);
-				trap->FS_Write(string, strlen(string), weaponfile);
+				trap->FS_Write(string, (int)strlen(string), weaponfile);
 			}
 
 			Com_sprintf(string, sizeof(string), "\n");
-			trap->FS_Write(string, strlen(string), weaponfile);
+			trap->FS_Write(string, (int)strlen(string), weaponfile);
 		}
 	}
 
 	// Sum up the totals.
 	Com_sprintf(string, sizeof(string), "\n***TOTAL:");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 	for (j=0;j<WP_NUM_WEAPONS;j++)
 	{
 		Com_sprintf(string, sizeof(string), "\t%d", totalpickups[j]);
-		trap->FS_Write(string, strlen(string), weaponfile);
+		trap->FS_Write(string, (int)strlen(string), weaponfile);
 	}
 
 	Com_sprintf(string, sizeof(string), "\n\n\n");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 
 	// Weapon fires per character
 	Com_sprintf(string, sizeof(string), "Weapon Shots per Player:\n\n");
-	trap->FS_Write( string, strlen( string ), weaponfile);
+	trap->FS_Write( string, (int)strlen( string ), weaponfile);
 
 	Com_sprintf(string, sizeof(string), "Player");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 	for (j=0; j<WP_NUM_WEAPONS; j++)
 	{
 		Com_sprintf(string, sizeof(string), "\t%s", weaponNameFromIndex[j]);
-		trap->FS_Write(string, strlen(string), weaponfile);
+		trap->FS_Write(string, (int)strlen(string), weaponfile);
 	}
 	Com_sprintf(string, sizeof(string), "\n");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 	// Cycle through each player, give their name and the number of times they picked up each weapon.
 	for (i=0; i<MAX_CLIENTS; i++)
@@ -446,47 +446,47 @@ void G_LogWeaponOutput(void)
 			{
 				nameptr = unknownname;
 			}
-			trap->FS_Write(nameptr, strlen(nameptr), weaponfile);
+			trap->FS_Write(nameptr, (int)strlen(nameptr), weaponfile);
 
 			for (j=0;j<WP_NUM_WEAPONS;j++)
 			{
 				Com_sprintf(string, sizeof(string), "\t%d", G_WeaponLogFired[i][j]);
-				trap->FS_Write(string, strlen(string), weaponfile);
+				trap->FS_Write(string, (int)strlen(string), weaponfile);
 			}
 
 			Com_sprintf(string, sizeof(string), "\n");
-			trap->FS_Write(string, strlen(string), weaponfile);
+			trap->FS_Write(string, (int)strlen(string), weaponfile);
 		}
 	}
 
 	// Sum up the totals.
 	Com_sprintf(string, sizeof(string), "\n***TOTAL:");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 	for (j=0;j<WP_NUM_WEAPONS;j++)
 	{
 		Com_sprintf(string, sizeof(string), "\t%d", totalshots[j]);
-		trap->FS_Write(string, strlen(string), weaponfile);
+		trap->FS_Write(string, (int)strlen(string), weaponfile);
 	}
 
 	Com_sprintf(string, sizeof(string), "\n\n\n");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 
 	// Weapon time per character
 	Com_sprintf(string, sizeof(string), "Weapon Use Time per Player:\n\n");
-	trap->FS_Write( string, strlen( string ), weaponfile);
+	trap->FS_Write( string, (int)strlen( string ), weaponfile);
 
 	Com_sprintf(string, sizeof(string), "Player");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 	for (j=0; j<WP_NUM_WEAPONS; j++)
 	{
 		Com_sprintf(string, sizeof(string), "\t%s", weaponNameFromIndex[j]);
-		trap->FS_Write(string, strlen(string), weaponfile);
+		trap->FS_Write(string, (int)strlen(string), weaponfile);
 	}
 	Com_sprintf(string, sizeof(string), "\n");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 	// Cycle through each player, give their name and the number of times they picked up each weapon.
 	for (i=0; i<MAX_CLIENTS; i++)
@@ -501,48 +501,48 @@ void G_LogWeaponOutput(void)
 			{
 				nameptr = unknownname;
 			}
-			trap->FS_Write(nameptr, strlen(nameptr), weaponfile);
+			trap->FS_Write(nameptr, (int)strlen(nameptr), weaponfile);
 
 			for (j=0;j<WP_NUM_WEAPONS;j++)
 			{
 				Com_sprintf(string, sizeof(string), "\t%d", G_WeaponLogTime[i][j]);
-				trap->FS_Write(string, strlen(string), weaponfile);
+				trap->FS_Write(string, (int)strlen(string), weaponfile);
 			}
 
 			Com_sprintf(string, sizeof(string), "\n");
-			trap->FS_Write(string, strlen(string), weaponfile);
+			trap->FS_Write(string, (int)strlen(string), weaponfile);
 		}
 	}
 
 	// Sum up the totals.
 	Com_sprintf(string, sizeof(string), "\n***TOTAL:");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 	for (j=0;j<WP_NUM_WEAPONS;j++)
 	{
 		Com_sprintf(string, sizeof(string), "\t%d", totaltime[j]);
-		trap->FS_Write(string, strlen(string), weaponfile);
+		trap->FS_Write(string, (int)strlen(string), weaponfile);
 	}
 
 	Com_sprintf(string, sizeof(string), "\n\n\n");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 
 
 	// Weapon deaths per character
 	Com_sprintf(string, sizeof(string), "Weapon Deaths per Player:\n\n");
-	trap->FS_Write( string, strlen( string ), weaponfile);
+	trap->FS_Write( string, (int)strlen( string ), weaponfile);
 
 	Com_sprintf(string, sizeof(string), "Player");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 	for (j=0; j<WP_NUM_WEAPONS; j++)
 	{
 		Com_sprintf(string, sizeof(string), "\t%s", weaponNameFromIndex[j]);
-		trap->FS_Write(string, strlen(string), weaponfile);
+		trap->FS_Write(string, (int)strlen(string), weaponfile);
 	}
 	Com_sprintf(string, sizeof(string), "\n");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 	// Cycle through each player, give their name and the number of times they picked up each weapon.
 	for (i=0; i<MAX_CLIENTS; i++)
@@ -557,31 +557,31 @@ void G_LogWeaponOutput(void)
 			{
 				nameptr = unknownname;
 			}
-			trap->FS_Write(nameptr, strlen(nameptr), weaponfile);
+			trap->FS_Write(nameptr, (int)strlen(nameptr), weaponfile);
 
 			for (j=0;j<WP_NUM_WEAPONS;j++)
 			{
 				Com_sprintf(string, sizeof(string), "\t%d", G_WeaponLogDeaths[i][j]);
-				trap->FS_Write(string, strlen(string), weaponfile);
+				trap->FS_Write(string, (int)strlen(string), weaponfile);
 			}
 
 			Com_sprintf(string, sizeof(string), "\n");
-			trap->FS_Write(string, strlen(string), weaponfile);
+			trap->FS_Write(string, (int)strlen(string), weaponfile);
 		}
 	}
 
 	// Sum up the totals.
 	Com_sprintf(string, sizeof(string), "\n***TOTAL:");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 	for (j=0;j<WP_NUM_WEAPONS;j++)
 	{
 		Com_sprintf(string, sizeof(string), "\t%d", totaldeaths[j]);
-		trap->FS_Write(string, strlen(string), weaponfile);
+		trap->FS_Write(string, (int)strlen(string), weaponfile);
 	}
 
 	Com_sprintf(string, sizeof(string), "\n\n\n");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 
 
@@ -589,18 +589,18 @@ void G_LogWeaponOutput(void)
 	// Weapon damage per character
 
 	Com_sprintf(string, sizeof(string), "Weapon Damage per Player:\n\n");
-	trap->FS_Write( string, strlen( string ), weaponfile);
+	trap->FS_Write( string, (int)strlen( string ), weaponfile);
 
 	Com_sprintf(string, sizeof(string), "Player");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 	for (j=0; j<WP_NUM_WEAPONS; j++)
 	{
 		Com_sprintf(string, sizeof(string), "\t%s", weaponNameFromIndex[j]);
-		trap->FS_Write(string, strlen(string), weaponfile);
+		trap->FS_Write(string, (int)strlen(string), weaponfile);
 	}
 	Com_sprintf(string, sizeof(string), "\n");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 	// Cycle through each player, give their name and the number of times they picked up each weapon.
 	for (i=0; i<MAX_CLIENTS; i++)
@@ -627,49 +627,49 @@ void G_LogWeaponOutput(void)
 			{
 				nameptr = unknownname;
 			}
-			trap->FS_Write(nameptr, strlen(nameptr), weaponfile);
+			trap->FS_Write(nameptr, (int)strlen(nameptr), weaponfile);
 
 			for (j=0;j<WP_NUM_WEAPONS;j++)
 			{
 				Com_sprintf(string, sizeof(string), "\t%d", percharacter[j]);
-				trap->FS_Write(string, strlen(string), weaponfile);
+				trap->FS_Write(string, (int)strlen(string), weaponfile);
 			}
 
 			Com_sprintf(string, sizeof(string), "\n");
-			trap->FS_Write(string, strlen(string), weaponfile);
+			trap->FS_Write(string, (int)strlen(string), weaponfile);
 		}
 	}
 
 	// Sum up the totals.
 	Com_sprintf(string, sizeof(string), "\n***TOTAL:");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 	for (j=0;j<WP_NUM_WEAPONS;j++)
 	{
 		Com_sprintf(string, sizeof(string), "\t%d", totaldamage[j]);
-		trap->FS_Write(string, strlen(string), weaponfile);
+		trap->FS_Write(string, (int)strlen(string), weaponfile);
 	}
 
 	Com_sprintf(string, sizeof(string), "\n\n\n");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 
 
 	// Weapon kills per character
 
 	Com_sprintf(string, sizeof(string), "Weapon Kills per Player:\n\n");
-	trap->FS_Write( string, strlen( string ), weaponfile);
+	trap->FS_Write( string, (int)strlen( string ), weaponfile);
 
 	Com_sprintf(string, sizeof(string), "Player");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 	for (j=0; j<WP_NUM_WEAPONS; j++)
 	{
 		Com_sprintf(string, sizeof(string), "\t%s", weaponNameFromIndex[j]);
-		trap->FS_Write(string, strlen(string), weaponfile);
+		trap->FS_Write(string, (int)strlen(string), weaponfile);
 	}
 	Com_sprintf(string, sizeof(string), "\n");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 	// Cycle through each player, give their name and the number of times they picked up each weapon.
 	for (i=0; i<MAX_CLIENTS; i++)
@@ -696,48 +696,48 @@ void G_LogWeaponOutput(void)
 			{
 				nameptr = unknownname;
 			}
-			trap->FS_Write(nameptr, strlen(nameptr), weaponfile);
+			trap->FS_Write(nameptr, (int)strlen(nameptr), weaponfile);
 
 			for (j=0;j<WP_NUM_WEAPONS;j++)
 			{
 				Com_sprintf(string, sizeof(string), "\t%d", percharacter[j]);
-				trap->FS_Write(string, strlen(string), weaponfile);
+				trap->FS_Write(string, (int)strlen(string), weaponfile);
 			}
 
 			Com_sprintf(string, sizeof(string), "\n");
-			trap->FS_Write(string, strlen(string), weaponfile);
+			trap->FS_Write(string, (int)strlen(string), weaponfile);
 		}
 	}
 
 	// Sum up the totals.
 	Com_sprintf(string, sizeof(string), "\n***TOTAL:");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 	for (j=0;j<WP_NUM_WEAPONS;j++)
 	{
 		Com_sprintf(string, sizeof(string), "\t%d", totalkills[j]);
-		trap->FS_Write(string, strlen(string), weaponfile);
+		trap->FS_Write(string, (int)strlen(string), weaponfile);
 	}
 
 	Com_sprintf(string, sizeof(string), "\n\n\n");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 
 
 	// Damage type damage per character
 	Com_sprintf(string, sizeof(string), "Typed Damage per Player:\n\n");
-	trap->FS_Write( string, strlen( string ), weaponfile);
+	trap->FS_Write( string, (int)strlen( string ), weaponfile);
 
 	Com_sprintf(string, sizeof(string), "Player");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 	for (j=0; j<MOD_MAX; j++)
 	{
 		Com_sprintf(string, sizeof(string), "\t%s", modNames[j]);
-		trap->FS_Write(string, strlen(string), weaponfile);
+		trap->FS_Write(string, (int)strlen(string), weaponfile);
 	}
 	Com_sprintf(string, sizeof(string), "\n");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 	// Cycle through each player, give their name and the number of times they picked up each weapon.
 	for (i=0; i<MAX_CLIENTS; i++)
@@ -752,48 +752,48 @@ void G_LogWeaponOutput(void)
 			{
 				nameptr = unknownname;
 			}
-			trap->FS_Write(nameptr, strlen(nameptr), weaponfile);
+			trap->FS_Write(nameptr, (int)strlen(nameptr), weaponfile);
 
 			for (j=0;j<MOD_MAX;j++)
 			{
 				Com_sprintf(string, sizeof(string), "\t%d", G_WeaponLogDamage[i][j]);
-				trap->FS_Write(string, strlen(string), weaponfile);
+				trap->FS_Write(string, (int)strlen(string), weaponfile);
 			}
 
 			Com_sprintf(string, sizeof(string), "\n");
-			trap->FS_Write(string, strlen(string), weaponfile);
+			trap->FS_Write(string, (int)strlen(string), weaponfile);
 		}
 	}
 
 	// Sum up the totals.
 	Com_sprintf(string, sizeof(string), "\n***TOTAL:");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 	for (j=0;j<MOD_MAX;j++)
 	{
 		Com_sprintf(string, sizeof(string), "\t%d", totaldamageMOD[j]);
-		trap->FS_Write(string, strlen(string), weaponfile);
+		trap->FS_Write(string, (int)strlen(string), weaponfile);
 	}
 
 	Com_sprintf(string, sizeof(string), "\n\n\n");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 
 
 	// Damage type kills per character
 	Com_sprintf(string, sizeof(string), "Damage-Typed Kills per Player:\n\n");
-	trap->FS_Write( string, strlen( string ), weaponfile);
+	trap->FS_Write( string, (int)strlen( string ), weaponfile);
 
 	Com_sprintf(string, sizeof(string), "Player");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 	for (j=0; j<MOD_MAX; j++)
 	{
 		Com_sprintf(string, sizeof(string), "\t%s", modNames[j]);
-		trap->FS_Write(string, strlen(string), weaponfile);
+		trap->FS_Write(string, (int)strlen(string), weaponfile);
 	}
 	Com_sprintf(string, sizeof(string), "\n");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 	// Cycle through each player, give their name and the number of times they picked up each weapon.
 	for (i=0; i<MAX_CLIENTS; i++)
@@ -808,31 +808,31 @@ void G_LogWeaponOutput(void)
 			{
 				nameptr = unknownname;
 			}
-			trap->FS_Write(nameptr, strlen(nameptr), weaponfile);
+			trap->FS_Write(nameptr, (int)strlen(nameptr), weaponfile);
 
 			for (j=0;j<MOD_MAX;j++)
 			{
 				Com_sprintf(string, sizeof(string), "\t%d", G_WeaponLogKills[i][j]);
-				trap->FS_Write(string, strlen(string), weaponfile);
+				trap->FS_Write(string, (int)strlen(string), weaponfile);
 			}
 
 			Com_sprintf(string, sizeof(string), "\n");
-			trap->FS_Write(string, strlen(string), weaponfile);
+			trap->FS_Write(string, (int)strlen(string), weaponfile);
 		}
 	}
 
 	// Sum up the totals.
 	Com_sprintf(string, sizeof(string), "\n***TOTAL:");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 	for (j=0;j<MOD_MAX;j++)
 	{
 		Com_sprintf(string, sizeof(string), "\t%d", totalkillsMOD[j]);
-		trap->FS_Write(string, strlen(string), weaponfile);
+		trap->FS_Write(string, (int)strlen(string), weaponfile);
 	}
 
 	Com_sprintf(string, sizeof(string), "\n\n\n");
-	trap->FS_Write(string, strlen(string), weaponfile);
+	trap->FS_Write(string, (int)strlen(string), weaponfile);
 
 
 	trap->FS_Close(weaponfile);
